third_party/WebKit/LayoutTests/inspector/console/console-log-linkify-stack-in-errors-expected.txt

CONSOLE MESSAGE: line 5: Error: Some test
    at namedFunction (foob.js:5:17)
    at foob.js:8:1
CONSOLE MESSAGE: line 11: Error: line
break
    at forStack (console-log-linkify-stack-in-errors.html:11:17)
    at console-log-linkify-stack-in-errors.html:14:1
CONSOLE MESSAGE: line 30: TypeError: Cannot read property 'removeChild' of null
    at domError (console-log-linkify-stack-in-errors.html:28:22)
    at console-log-linkify-stack-in-errors.html:34:1
CONSOLE MESSAGE: line 41: Error: some error
CONSOLE MESSAGE: line 47: Error message without stacks http://www.chromium.org/
CONSOLE MESSAGE: line 49: Error valid stack #2
    at http://www.chromium.org/boo.js:40:70
    at foo(http://www.chromium.org/foo.js:10:50)
CONSOLE MESSAGE: line 50: Error valid stack #3
    at http://www.chromium.org/foo.js:40
CONSOLE MESSAGE: line 51: Error: MyError
    at throwError (http://www.chromium.org/foo.js:40)
    at eval (eval at <anonymous> (http://www.chromium.org/foo.js:42:1), <anonymous>:1:1)
    at http://www.chromium.org/foo.js:239
CONSOLE MESSAGE: line 20: ReferenceError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20:21)
    at stack1 (console-log-linkify-stack-in-errors.html:22:5)
    at console-log-linkify-stack-in-errors.html:53:1
CONSOLE MESSAGE: line 20: EvalError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20:21)
    at stack1 (console-log-linkify-stack-in-errors.html:22:5)
    at console-log-linkify-stack-in-errors.html:54:1
CONSOLE MESSAGE: line 20: SyntaxError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20:21)
    at stack1 (console-log-linkify-stack-in-errors.html:22:5)
    at console-log-linkify-stack-in-errors.html:55:1
CONSOLE MESSAGE: line 20: RangeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20:21)
    at stack1 (console-log-linkify-stack-in-errors.html:22:5)
    at console-log-linkify-stack-in-errors.html:56:1
CONSOLE MESSAGE: line 20: TypeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20:21)
    at stack1 (console-log-linkify-stack-in-errors.html:22:5)
    at console-log-linkify-stack-in-errors.html:57:1
CONSOLE MESSAGE: line 20: URIError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:20:21)
    at stack1 (console-log-linkify-stack-in-errors.html:22:5)
    at console-log-linkify-stack-in-errors.html:58:1
CONSOLE MESSAGE: line 60: Error broken stack
    at function_name(foob.js foob.js:30:1)
 at foob.js:40:70
CONSOLE MESSAGE: line 61: Error broken stack #2
    at function_name(foob.js:20:30
CONSOLE MESSAGE: line 62: Error broken stack #3
    at function_name(foob:20.js:30   bla
CONSOLE MESSAGE: line 63: Error broken stack #4
    at function_name)foob.js:20:30(
CONSOLE MESSAGE: line 64: Error broken stack #5
    at function_name foob.js:20:30)
CONSOLE MESSAGE: line 65: Error broken stack #6
    at foob.js foob.js:40:70
CONSOLE MESSAGE: line 9: Error
    at Failure.letsFailWithStack (stack-with-sourceMap.js:19:20)
    at <anonymous>:1:1
Test that console.log(new Error().stack) would linkify links in stacks for sourceUrls and sourceMaps Bug 424001.

foob.js:5 Error: Some test
    at namedFunction (foob.js:5)
    at foob.js:8 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:4 Error: line
break
    at forStack (console-log-linkify-stack-in-errors.html:4)
    at console-log-linkify-stack-in-errors.html:7 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:23 TypeError: Cannot read property 'removeChild' of null
    at domError (console-log-linkify-stack-in-errors.html:21)
    at console-log-linkify-stack-in-errors.html:27 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:34 Error: some error
    at logError (console-log-linkify-stack-in-errors.html:32)
    at console-log-linkify-stack-in-errors.html:38 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > object-value-error source-code > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:40 Error message without stacks http://www.chromium.org/ console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:42 Error valid stack #2
    at www.chromium.org/boo.js:40
    at foo(www.chromium.org/foo.js:10) console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:43 Error valid stack #3
    at www.chromium.org/foo.js:40 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:44 Error: MyError
    at throwError (www.chromium.org/foo.js:40)
    at eval (eval at <anonymous> (www.chromium.org/foo.js:42), <anonymous>:1:1)
    at www.chromium.org/foo.js:239 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:13 ReferenceError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:13)
    at stack1 (console-log-linkify-stack-in-errors.html:15)
    at console-log-linkify-stack-in-errors.html:46 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:13 EvalError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:13)
    at stack1 (console-log-linkify-stack-in-errors.html:15)
    at console-log-linkify-stack-in-errors.html:47 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:13 SyntaxError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:13)
    at stack1 (console-log-linkify-stack-in-errors.html:15)
    at console-log-linkify-stack-in-errors.html:48 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:13 RangeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:13)
    at stack1 (console-log-linkify-stack-in-errors.html:15)
    at console-log-linkify-stack-in-errors.html:49 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:13 TypeError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:13)
    at stack1 (console-log-linkify-stack-in-errors.html:15)
    at console-log-linkify-stack-in-errors.html:50 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:13 URIError: valid stack
    at stack2 (console-log-linkify-stack-in-errors.html:13)
    at stack1 (console-log-linkify-stack-in-errors.html:15)
    at console-log-linkify-stack-in-errors.html:51 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link > webkit-html-resource-link
console-log-linkify-stack-in-errors.html:53 Error broken stack
    at function_name(foob.js foob.js:30:1)
 at foob.js:40:70 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text
console-log-linkify-stack-in-errors.html:54 Error broken stack #2
    at function_name(foob.js:20:30 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text
console-log-linkify-stack-in-errors.html:55 Error broken stack #3
    at function_name(foob:20.js:30   bla console-message > source-code > console-message-url webkit-html-resource-link > console-message-text
console-log-linkify-stack-in-errors.html:56 Error broken stack #4
    at function_name)foob.js:20:30( console-message > source-code > console-message-url webkit-html-resource-link > console-message-text
console-log-linkify-stack-in-errors.html:57 Error broken stack #5
    at function_name foob.js:20:30) console-message > source-code > console-message-url webkit-html-resource-link > console-message-text
console-log-linkify-stack-in-errors.html:58 Error broken stack #6
    at foob.js foob.js:40:70 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text
stack-with-sourceMap.coffee:3 Error
    at Failure.letsFailWithStack (stack-with-sourceMap.coffee:3)
    at window.failure (stack-with-sourceMap.coffee:7)
    at <anonymous>:1:1 console-message > source-code > console-message-url webkit-html-resource-link > console-message-text > webkit-html-resource-link > webkit-html-resource-link
